Isolate most tests
