+++ tests/fabric.chain.js
@@ -1,0 +1,69 @@
+'use strict';
+
+const Fabric = require('../');
+const assert = require('assert');
+
+describe('@fabric/core/types/chain', function () {
+  describe('Chain', function () {
+    it('is available from @fabric/core', function () {
+      assert.equal(Fabric.Chain instanceof Function, true);
+    });
+
+    it('can cleanly start and stop a chain', async function () {
+      let chain = new Fabric.Chain();
+
+      await chain.start();
+      await chain.stop();
+
+      assert.ok(chain);
+      assert.ok(chain.ledger);
+    });
+
+    it('can append an arbitrary message', async function () {
+      let chain = new Fabric.Chain();
+
+      await chain.start();
+      await chain.append({ debug: true, input: 'Hello, world.' });
+      await chain.stop();
+
+      assert.ok(chain);
+      assert.ok(chain.ledger);
+    });
+
+    xit('generates a merkle tree with the expected proof of inclusion', async function () {
+      let chain = new Fabric.Chain();
+
+      await chain.start();
+      await chain.append({ debug: true, input: 'Hello, world.' });
+      await chain.append({ debug: true, input: 'Why trust?  Verify.' });
+      await chain.stop();
+
+      let sample = chain.blocks.map(b => Buffer.from(b['@id'], 'hex'));
+      let tree = chain['@tree'];
+      let root = tree.getRoot();
+
+      let proofs = {
+        genesis: tree.getProof(sample[0], 0),
+        'blocks/1': tree.getProof(sample[1], 1),
+        'blocks/2': tree.getProof(sample[2], 2)
+      };
+
+      let verifiers = {
+        genesis: tree.verify(proofs.genesis, sample[0], root),
+        'blocks/1': tree.verify(proofs['blocks/1'], sample[1], root),
+        'blocks/2': tree.verify(proofs['blocks/2'], sample[2], root),
+        invalid: tree.verify(proofs['genesis'], Buffer.alloc(32), root)
+      };
+
+      assert.ok(chain);
+      assert.equal(sample.length, 3);
+      assert.equal(sample[0].toString('hex'), 'c1b294376d6d30d85a81cff9244e7b447a02e6307a047c4a53643a945022e505');
+      assert.equal(sample[1].toString('hex'), '67822dac02f2c1ae1e202d8e75437eaede631861e60340b2fbb258cdb75780f3');
+      assert.equal(sample[2].toString('hex'), 'a59402c14784e1be43b1adfc7832fa8c402dddf1ede7f7c29549d499b112444f');
+      assert.equal(verifiers.genesis, true);
+      assert.equal(verifiers['blocks/1'], true);
+      assert.equal(verifiers['blocks/2'], true);
+      assert.equal(verifiers.invalid, false);
+    });
+  });
+});
