+++ tests/fabric.machine.js
@@ -1,0 +1,68 @@
+'use strict';
+
+const Fabric = require('../');
+const assert = require('assert');
+
+const OPCODES = require('../assets/opcodes');
+
+describe('@fabric/core/types/machine', function () {
+  describe('Machine', function () {
+    it('is available from @fabric/core', function () {
+      assert.equal(Fabric.Machine instanceof Function, true);
+    });
+
+    it('can compute a value', async function prove () {
+      // TODO: use Fabric itself
+      let machine = new Fabric.Machine(false);
+
+      // TODO: use Fabric instead of Fabric.Machine
+      machine.define('OP_TRUE', OPCODES.OP_TRUE);
+
+      // fabric.push('OP_TRUE');
+      machine.script.push('OP_TRUE');
+
+      await machine.start();
+      await machine.compute();
+      await machine.stop();
+
+      // console.log('machine state:', machine.state);
+
+      // assert.equal(machine.state.id, samples.names.encodedStackWithSingleValidFrame);
+      assert.equal(machine.state['@data'][0], true);
+    });
+
+    it('can correctly sum two values', async function prove () {
+      let machine = new Fabric.Machine(false);
+
+      machine.define('OP_ADD', OPCODES.OP_ADD);
+
+      machine.script.push('1');
+      machine.script.push('1');
+      machine.script.push('OP_ADD');
+
+      await machine.start();
+      await machine.compute();
+      await machine.stop();
+
+      assert.equal(machine.state['@data'][0], 2);
+    });
+
+    it('can correctly sum three values', async function prove () {
+      let machine = new Fabric.Machine(false);
+
+      machine.define('OP_ADD', OPCODES.OP_ADD);
+
+      machine.script.push('1');
+      machine.script.push('1');
+      machine.script.push('OP_ADD');
+      machine.script.push('2');
+      machine.script.push('OP_ADD');
+
+      await machine.start();
+      await machine.compute();
+      await machine.stop();
+
+      assert.equal(machine.state['@data'][0], 4);
+    });
+  });
+});
